# docs/api-specification.yaml
openapi: 3.0.0
info:
  title: EduAble Platform API
  description: Comprehensive educational platform for coaching institutes
  version: 1.0.0
  contact:
    name: EduAble API Support
    email: api-support@eduable.com
  license:
    name: Proprietary
    url: https://eduable.com/terms
servers:
  - url: https://api.eduable.com/v1
    description: Production server
  - url: https://staging-api.eduable.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
        '429':
          description: Rate limit exceeded
  /auth/register:
    post:
      tags: [Authentication]
      summary: User registration
      description: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Validation error
        '409':
          description: User already exists
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: student@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: "securePassword123"
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [student, teacher, admin, institute_admin]
        instituteId:
          type: string
        avatar:
          type: string
          format: uri
...